+++ hint.go
@@ -2,6 +2,13 @@ package readline
 
 import "regexp"
 
+// SetHintText - a nasty function to force writing a new hint text. It does not update helpers, it just renders
+// them, so the hint will survive until the helpers (thus including the hint) will be updated/recomputed.
+func (rl *Instance) SetHintText(s string) {
+	rl.hintText = []rune(s)
+	rl.renderHelpers()
+}
+
 func (rl *Instance) getHintText() {
 
 	if !rl.modeAutoFind && !rl.modeTabFind {
@@ -17,6 +24,7 @@ func (rl *Instance) getHintText() {
 	}
 }
 
+// writeHintText - only writes the hint text and computes its offsets.
 func (rl *Instance) writeHintText() {
 	if len(rl.hintText) == 0 {
 		rl.hintY = 0
@@ -25,6 +33,8 @@ func (rl *Instance) writeHintText() {
 
 	width := GetTermWidth()
 
+	// Wraps the line, and counts the number of newlines in the string,
+	// adjusting the offset as well.
 	re := regexp.MustCompile(`\r?\n`)
 	newlines := re.Split(string(rl.hintText), -1)
 	offset := len(newlines)